requirements.txt
Flask
pandas
numpy
scikit-learn
mysql-connector-python
